refac(91026): Extrai task gerar_notificacao_atraso_entrega_ajustes_prestacao_de_contas_async
