+++ sme_ptrf_apps/core/tasks/__init__.py
@@ -9,5 +9,6 @@ from .gerar_notificacao_inicio_periodo_pc import gerar_notificacao_inicio_period
 from .gerar_notificacao_pendencia_envio_pc import gerar_notificacao_pendencia_envio_prestacao_de_contas_async
 from .gerar_notificacao_proximidade_inicio_pc import gerar_notificacao_proximidade_inicio_periodo_prestacao_conta_async
 from .gerar_notificacao_proximidade_fim_periodo_pc import gerar_notificacao_proximidade_fim_periodo_prestacao_conta_async
+from .gerar_notificacao_atraso_entrega_ajustes_pc import gerar_notificacao_atraso_entrega_ajustes_prestacao_de_contas_async
